refactor(core): use function programming to build `getEmbeds`
Remove the imperative `getEmbeds` implementation and replace it by a
functional implementation. The goal of this exercise is to move towards
barrel files that contain no "new implementations", but only
compositions of (unit-tested) functions.

Correspondingly, all tests focussing on the `getEmbeds` function have
become integration testing (notably, this reduces the duplication
between the unit and integration test suite for the embeds folder). As a
consequence, the integration test suite covers nearly 100% of the embeds
logic and the unit tests everything but the barrel file (excluding the
`functional.ts` module, which will be tested separately).

The functions required for this, found in `functional.ts` are currently
included with the embeds logic but should be either moved to a top-level
module of core, a new package, or replaced by a pre-existing package. A
pre-existing package was skipped here both as a learning exercise and
due to type checking issues with the packages that were tried.