chore(scripts): improve calls to linter CLIs
Following the documentation changes for markdownlint-cli 0.32.1->0.32.2,
this adds "--" between the list of options/flags for linter CLIs (ESLint
and markdownlint-cli) to separate it from the arguments listing the
files that should be linted.

This prevents any potential issues with "variadic" arguments (which
accept multiple values). It was chosen to always add "--  to avoid
forgetting to do it for any specific linter when required in the future
and since using "--" does not change the behaviour of the CLI.