refactor(core): expose `extract.ts` internals for testability
Expose the internal functions used in the `extract.ts` file to be able
to test them directly, not indirectly through whatever functions are
exposed. This allows for more precise testing and easier coverage of
edge case.

The signature of `compareStartIndex` was simplified to avoid having to
construct complete instances of `WebManglerEmbed` for the purposes of
testing something that relates to only a single character. This also
simplifies the signature for users of the function, which is good in
general, e.g. for re-usability.

A next step is to compose the exposed functionality dynamically rather
than implicitly composing the functions prior to export (aka dependency
injection).