refactor(mangler-css-variables): convert re-as-string to /regex/.source
Convert all strings in the package the represent a Regular Expression
into a regular expression, and convert it back to a string immediately
using the `.source` property.

Having these regular expressions as literals enable static analysis
tools (currently at least ESLint and CodeQL) to process them and flag
potential problems. It also makes it more explicit that these are
regular expressions, and enable code highlighting on them.

Also align the tests with minor adjustments to the regular expressions,
notably removing an unnecessary escape character and making a group
non-capturing.

Lastly, fix some indentations for `interface`s in alignment with ESLint.