refactor(core): use dependency injection for `extractEmbedsFromContent`
Create a builder function for the `extractEmbedsFromContent` that takes
the previously exposed helper functions to build the function with
explicit dependencies. This will allow for removing assumptions &
estimates from the test, such as those related to embed ids.

The test suite has only been updated to make them run. As the unit test
suite for this function now is, explicitly, an integration test, the
suite has to be updated accordingly. This is left as a separate task to
reduce the scope of this commit.