feat(scripts): determine vcs changes based on repo state
Update the VCS helpers to determine "what changed" based on the repo
state. If the repo is dirty, testing and linting will only apply to the
dirty part of the repo, rather than everything that changed relative to
`main`. When the repo isn't dirty, "what changed" is still everything
that changed compared to `main`.