chore(deps): bump Node.js (for development) from 16 to 18
Bump the recommended Node.js version for development 16 to 18.
Specifically, the latest Node.js version currently (stable) available is
specified in the .nvmrc file. And the `engines` field of the root
package.json is updated to require at a minimum Node.js 18.0.0 when
running npm commands (such as `npm install`).