fix: compiler errors due to type changes
Fix errors from the compiler in other packages due to changes to the
`QuerySelectorOptions` type in the types package.

For the languages plugins this comes down to removing some code that
hasn't been released yet.

For the mangler plugins this comes down to adding a value for the `kind`
field where necessary, which currently has no further effect.