test(core): improve unit test for `extractEmbedsFromContent`
First, uncouple the `extractEmbedsFromContent` test suite from the
`generateUniqueString` implementation. This way, a failed test in that
suite is guaranteed not to be caused by a wrong implementation of the
`generateUniqueString` function. Moreover, it further decouples the test
from knowledge about the system by making it independent of the
`idCharSet`.

Additionally, the sorting function has actually been removed as an
explicit dependency of `extractEmbedsFromContent`. This is because the
sorting is pretty much invisible from the outside. For this reason it's
hard to provide a mock/stub for the test. Hence, the function will be
considered an implementation detail of `extractEmbedsFromContent`.