test(core): improve unit test for `generateUniqueString`
Simplify the unit test suite for the `generateUniqueString` function by
making the `CharSet` to generate substrings from an input to a builder
function that returns a `generateUniqueString` function. This way, the
test doesn't need to have knowledge about a `CharSet` somewhere in the
code base.

The alternative, providing a `NameGenerator` to the builder, was
considered but not chosen as, due to the state contains, this would lead
to suboptimal string generator on repeated use. Namely, if the first
call to `generateUniqueString` returns the unique string "aab" then the
next call will only generate strings that come after "aab", even though
a shorter string might work for the string provided.