config: enable first eslint-plugin-regexp rule
Configure the `eslint-plugin-regexp` as a plugin in the ESLint
configuration and enable the first rule of the plugin. The rule is the
`match-any` rule [1], which enforces the use of a consistent pattern to
match any string. It is configured to allow using `/[^]/` (match
absolutely everything) and `/./` (match everything except newlines). The
latter is named "dotAll" in the configuration.

This rule is applied throughout the code base where the pattern
`/(?:.|\n|\r)/` was used to match "absolutely everything".

--
1. https://ota-meshi.github.io/eslint-plugin-regexp/rules/match-any.html