refactor(language-css): pass language options to all helpers
Update the helpers interface to require passing the options parameter to
all helper functions. This will simplify changes when these helpers do
need the options at a point in the future - or in the case of languages,
already now by avoiding positional arguments.

Also took the opportunity to, while adjusting the helpers test suite,
update the `TestScenario` type to use `readonly` fields to avoid the
test changes values unintentionally.