feat(core): force read-only on custom plugin options
Update the options interfaces for all plugins in the core that bundle
multiple plugins to use only `readonly` properties. Note that the
interface for the options of the `BuiltInLanguagesPlugin` doesn't
require any changes as it has no (locally-defined) properties.

Also, per the 'typescript-eslint/eslint-plugin' rule 'consistent-type-
exports', updated the the export statements in all changed files.