refactor(core): use base58 character set for internal usage
The package core has some internal uses for character sets, primarily
for intermediary values and temporary identifiers. This introduces a
base58 character set (tests included) for use in such scenarios. It also
updates all internal character sets by this new character set.

Using the base58 character set for this should help for things like
debugging. Given the nature of this character set is to avoid mistaking
certain characters for others, it should improve the readability of the
affected internal values - which are seen primarily when debugging.