chore(types): remove field from `CssDeclarationValueOptions`
Remove the `caseSensitive` field from the `CssDeclarationValueOptions`
type. This follows reconsiderations [1] to the same fields on the
`QuerySelectorOptions` - more specifically the fact that the case
insensitivity can ultimately be language dependent. At the time of this
commit no alternative solution to handling case insensitivity of (some)
CSS values has been decided on (the primary goal is instead to avoid
releasing the reverted [2] change).

This is NOT a breaking change as the introduction of the removed field
hasn't been released.

--
1. Manifested in 93d38cf7e65be9f5c155772119fbe68d6c306404
2. Reverting d70108e7e59953ac705bbbdb25011aa7fee409bd