chore(language-js): start `kind` support for query selector expressions
Rewrite the `querySelectorExpressionFactory` to start support for the
`kind` field of the `QuerySelectorOptions`. This entails calling the old
implementation as a fallback when `kind` isn't set, and running a (to be
implemented) new implementation when it is.

The test suite for this factory is updated to, on top of what it
previously tested, test that the old implementation is called when
`kind` isn't set. Tests for when `kind` is set are to be added.