config: enable 'switch-exhaustiveness-check' rule in ESLint
Update the ESLint config to enable the 'switch-exhaustiveness-check'
rule from the typescript-eslint/eslint-plugin. This rule enforces that
all `switch`-statements are exhaustive, meaning there's no value that is
allowed by the type system that match one of the branches. This can be
helpful to prevent bugs.

Also, update the link to the rules documentation from the old repo-based
URL (now invalid) to the new website-based URL.