test(language-js): the `kind` implementation of query selector factory
Update both the unit and integration test suite to include a test suite
to test the expected behavior of the implementation for the `kind`
option of the query selector expressions factory.

The new unit test closely mimics the old ones, just with new inputs. The
integration tests mimic the old ones even closer, essentially just
removing the `suffix` and `prefix` options.

In both cases, the old test suite is kept. This is done through a second
top-level suite to limit the size of the diff (at the cost of needing an
eslint-ignore comment).