config: fully and explicitly configure eslint-plugin-mocha
Update the ESLint configuration to fully and explicitly configure the
eslint-plugin-mocha plugin rather than relying on extending
`plugin:mocha/recommended` (which has been removed).

This helps make the intent of the configuration clearer, makes it easier
to track changes over time, and makes it easier to change the config if
needed.

Most rules are enabled because they were previously enabled by the
recommended config and/or because they just make sense.

Two rules are set to "warn", these both relate to tests that are there
but aren't run. These are set to warn because it can be useful to have
such "tests". Skipped tests can be useful if a test is failing due to a
dependency. Unimplemented tests can be useful to convey the intend to
test something later. In both cases, human judgement is necessary to
determine if the test is acceptable.

Some rules were disabled, the reasons being:

- `mocha/no-empty-description`: disabled because this is implied by the
  set of valid-* rules. Running fewer rules should improve performance.
- `mocha/no-hooks`: We want to use hooks for, e.g., common setup.
- `mocha/no-hooks-for-single-case`: Having a hook for a single test can
  be fine. E.g. if more tests are expected or other tests were removed.
- `mocha/no-setup-in-describe`: Disabled because this project uses
  generated tests. (This was already disabled, but it was moved).
- `mocha/no-sibling-hooks`: Hooks may be separated by topic, e.g. one
  hook to set up stubs and one hook to set up data. This may lead to
  shorter hooks, which in turn are easier to reason about.
- `mocha/no-synchronous-tests`: Most tests are synchronous, so enabling
  this rules doesn't make sense.

Lastly, the rule `prefer-arrow-callback` is disabled for tests because
the rule `mocha/prefer-arrow-callback` is enabled (for details, see the
docs of the latter rule).