config: fully and explicitly configure eslint-plugin-chai-expect
Update the ESLint configuration to fully and explicitly configure the
eslint-plugin-chai-expect plugin rather than relying on extending
`plugin:chai-expect/recommended` (which has been removed).

This helps make the intent of the configuration clearer, makes it easier
to track changes over time, and makes it easier to change the config if
needed.

Like the recommended configuration, all rules are enabled. Note that the
`terminating-properties` rule is configured with the properties from
the sinon-chai plugin based on it's documentation [^1].

--
^1: See the "Assertions" section of the README.md in commit `e23406f`.