config: organize non-chai/mocha ESLint rules for test files
Organize the rules for test files that aren't directly related to test
files specifically (i.e. chai or mocha related). Removing 1 unnecessary
rule and re-ordering 1 disabled rule.

The disabled rule is `import/no-extraneous-dependencies`. Per the commit
message of 224a48365166e31ae1dd0f6c8c1836d644810c43 its disabled because

> Test files are allowed to [have] extraneous dependencies because their
> dependencies may come from the monorepo project rather than the local
> project. For example, chai [and sinon].

It was moved so as to group disabled rules to the start of the overrides
to make them more visible (plus, grouping by plugin makes less sense if
only at most a handful or rules are to be configured).

The rule `security/detect-non-literal-fs-filename` was removed because
the project currently has no violations for this rule. In general, we
don't want to be reading files in tests anyway.