feat(scripts): instruct ESLint to report unused disable directives
Following [1], this adds the `--report-unused-disable-directives` flag
when running ESLint. This ensures that ESLint fails and reports on
useless `eslint-disable` comments.

The flag was chosen over adding it to the ESLint config file because the
former makes ESLint fail (nonzero exit code) whereas the latter doesn't.
We want to hard fail as useless ignore directives should always be
removed.

--
1. https://eslint.org/docs/latest/user-guide/configuring/rules