config: enable Stryker's incremental mode
Update the Stryker configuration to enable incremental mode. This is
achieved by setting `incremental` to true and specifying a location to
store the incremental data file. The file is named based on the
package(s) being mutation tested. As such, rerunning mutation tests for
the same (set of) package(s) will utilize the incremental data file if
present from a previous run.

The data file is stored in a new generated folder called "cache", so
named because it caches data to improve the speed of, in this case,
mutation testing. This directory is configured to be ignored by git and
defined in `.values.cjs` to enable reuse of the cache dir in other
scripts and configurations. For example, the clean script already uses
it to remove the cache dir in `--hard` mode.