test(language-css): improve expression integration tests performance
Improve the performance of the integration tests for the language-css
package by reducing the number of test cases generated per scenario. In
particular, this is achieved by cutting back on the number of value sets
that are not adjacent to what is being tested and are therefor less
likely to influence the scenario.

Note that as part of these changes, some scenarios did actually see an
increase in the amount of test cases. This is purely because upon
evaluation of the scenario, it was deemed necessary to add the cases in
question.