ci: cache Stryker's incremental data
Cache the incremental data for mutation testing from Stryker to speed up
mutation testing in the CI. The use of the cache will ensure the
incremental data from a previous run (if available) will be loaded into
the correct file location before mutation testing beings. The specific
cache key chosen, i.e. including `${{ github.run_number }}`, ensures the
incremental data in the cache is kept up-to-date.

Because mutation testing now relies on two caches, the cache key for the
compiled packages cache has been updated to ensure the two caches are
kept separate. The caches are intentionally not combined as they serve
distinct purposes - removing one can be done safely without it affecting
the other - plus the incremental data is package specific in opposition
to the compiled packages.