test(language-css): expand end-2-end suite
Expand the end-to-end (e2e) test suite for the language-css package by
adding some basic tests for mangling specific kinds of things. The tests
are based on the existing expressions that exist in the plugin.

Note that one test is disabled because of a bug discovered as part of
this testing effort.

To make it more clear from the file name what's being tested, the
pre-existing test e2e test file was renamed.