config: divide coverage reports by package
Update the nyc configuration to output coverage reports into a subfolder
of `_reports/coverage` based on what package was tested. This separation
helps determine for what package a coverage report is without having to
open it, as well as an easier time switching between and comparing
coverage reports for different packages.

To avoid long folder names (which can be a problem on some file systems)
when testing multiple or all packages the placeholder name "_mixed" is
used in such cases. The leading "_" was chosen to prevent a potential
collision with a packages named "mixed" (it is not expected that package
names will ever start with an underscore).

The same logic is applied to the nyc's temporary data. While this will
have less impact on day-to-day development, having the separation may
still proof useful.