config: explicitly configure eslint-plugin-security
Remove the extension of `plugin:security/recommended` and instead
explicitly configure all rules of the plugin. Making the configuration
explicit makes it easier to work with.

Rules are enabled as much as possible. What follows is motivations for
why some rules are disabled:
- `detect-disable-mustache-escape`: This rule is about frontend/web
  based code, so not applicable to this project.
- `detect-no-csrf-before-method-override`: This rule is about server
  side code, so not applicable to this project.
- `security/detect-non-literal-regexp`: non-literal regular expression
  are used extensively and intentionally. It's not considered a security
  risk in the context of this project because regular expression are
  used for configuration and intended to run in a trusted environment.
- `security/detect-object-injection`: disabled due to the number of
  reports. This rule may be enabled in the future if a good alternative
  is found.
- `security/detect-unsafe-regex`: may be enabled later, possibly in
  conjunction with security related rules from `eslint-plugin-regexp`.

Note that for scripts one of the rules has been enabled, namely the
child_process rule. The motivation being two-fold. First, there are
currently no violations so that's a good indication it's not too noisy.
Second, in practice all scripts _should_ use `utilities/exec.js`, which
in turn should be the only file for which the rule can/has to be
disabled.