config: separate mutation reports by package
Update the Stryker configuration to output coverage reports into a
subfolder of `_reports/mutation` based on what package was tested. This
separation helps determine for what package a coverage report is without
having to open it, as well as an easier time switching between and
comparing coverage reports for different packages.

To avoid long folder names (which can be a problem on some file systems)
when testing multiple or all packages the placeholder name "_mixed" is
used in such cases. The leading "_" was chosen to prevent a potential
collision with a packages named "mixed" (it is not expected that package
names will ever start with an underscore).

Similar logic is already used for Stryker's incremental data. It's not
applied to the temporary data since Stryker copies the whole project
(all packages) anyway.