ci: create workflow to scan for secrets
Create a new GitHub Actions workflow called "Audit" to audit the
project, for now only scanning for secrets. The secret scan is done
using gitleaks [1], specifically the gitleaks-action [2]. This tool
is used because it's open source and free to use for open source.

The gitleaks-action is configured to not have any external effects.
The workflow is configured to run on all Pull Request and on the `main`
branch (like other workflows) as well as once a week on a schedule. The
latter will perform a historic scan, this is particularly useful the if
gitleaks is updated to detect new secrets.

--
1. https://github.com/zricethezav/gitleaks
2. https://github.com/gitleaks/gitleaks-action