config: update inclusion of eslint-plugin-import plugin
Update how eslint-plugin-import is included in ESLint by removing the
`plugin:import/recommended` extension and adding it as a "plugin"
directly instead. This aligns the configuration with the fact that all
rules are manually configured.

Note that `plugin:import/typescript` is kept as an extension as it's
truly about extending this configuration to support TypeScript imports
and exports.