config: disable `max-len` rule for YAML files
Disable the `max-len` rule [1] for YAML files as there is a violation
without an immediate solution in the code-analysis.yml workflow (line
31).

Ideally the maximum line length of 80 (or optionally a higher value for
YAML files) is enforce with an exception for continuous string values.
However, as of this commit I'm not aware of a way to configure this
(with the ESLint plugin currently in the project).

--
1. https://eslint.org/docs/latest/rules/max-len