test(language-js): improve performance of tests for expression
Improve the performance of the integration tests for the language-js
package by reducing the number of test cases generated per scenario. In
particular, this is achieved by cutting back on the number of value sets
that are not adjacent to what is being tested and are therefor less
likely to influence the outcome of the test.