test(language-css): better attributes with or without quotes
Update both the end-to-end (e2e) and integration tests for the
language-css package to better capture single quoted and unquoted
values.

The e2e tests were updated to include an example with single quotes and
without quotes for both the "multi-value-attributes" and "single-value-
attributes" test cases.

The integration tests were updated by means of updating the attribute
selector builder helper function to include unquoted examples if the
provided value does not include spaces.