config: configure 26 rules for eslint-plugin-regexp
Update the ESLint configuration to enable the (alphabetically) first 27
rules of the eslint-plugin-regexp (1 was already configured). By and
large this just enables the rules in an effort to standardize and
improve regular expressions throughout the project. If rules allow for
configuration it was attempted to match the current project's style.

Some things worth pointing out:
- `no-dupe-disjunctions`: is set to report as much as possible to reduce
  the chances of (potentially) bad regular expressions.
- `no-lazy-ends`: is set to report as much as possible to reduce the
  chances of (potentially) bad regular expressions.
- `no-misleading-unicode-character`: is set to be not-fixable as the fix
  algorithm makes assumptions. Manual review is preferred.
- `no-obscure-range`: is configured conservatively. When it makes sense,
  more ranges can be added on demand.

Also, update the link to the documentation to point to the website
(instead of GitHub README) as this is where the documentation is most
accessible.