fix: compiler errors due to type changes
Fix errors from the compiler in other packages due to changes to the
`CssDeclarationValueOptions` type in the types package.

For language plugins this comes down to adding a value for the `kind`
field where necessary in tests. For mangler plugins this comes down to
adding a value for the `kind` field where necessary, which currently
has no further effect.