chore(deps): add licensee
Add licensee [1] as a new development dependency, to be used for license
validation instead of Fossa (used through a GitHub integration at the
moment).

Using licensee has a couple of advantages over Fossa, most notably:
1. Validation can be performed offline
2. The license policy is version controlled

licensee specifically was chosen because it's conveniently available as
an npm package and works well for npm-only projects.

--
1. https://www.npmjs.com/package/licensee