config: create `license-check` script and configure licensee
Create a new npm script called `license-check` to run licensee and check
the licenses of all dependencies against the in-this-commit-added
license policy.

The license policy can be found in the `.licensee.json` file, the config
file for licensee [1]. The policy is found at `licenses.spdx`, which
lists the allowed licenses. The `corrections` option is set to true to
leverage crowdsourced corrections for packages that incorrectly
specified their license - more details can be found in the licensee
README.

--
1. https://www.npmjs.com/package/licensee