config: update mutation testing target thresholds
Update the thresholds for mutation testing (percentage of surviving
mutants) to break at 90% instead of 50%, significantly raising the bar.
As some packages do not yet meet this threshold, they've been added as
exceptions.

The aim of this change is to keep packages with good test suites up to a
higher standard and gradually improve the test suites of the other
packages over time.

This change is primarily aimed at individual packages, at this point in
time we're less interested in the overall score or combined scores.