config: add test-type specific coverage exclusions
Update the NYC configuration to use exclusions depending on the test
type (e.g. unit or integration). This is to get a more representative
coverage report as certain parts of the code may not be targeted by a
given test type.

In particular, this adds exclusions for integration and end-to-end
tests, neither of which intent to cover the __tests__ dirs (which
contain test helpers and (unit) tests for those helpers).