ci: run unit and integration tests separately
Split the test job in the continuous integration setup into one for unit
tests and one for integration tests. This should improve the pipeline by
making it clearer what went wrong (if anything), and failing earlier if
there's a problem detectable by unit tests. Additionally, it allows for
running the mutation tests earlier as those only depend on unit tests.

This also means coverage reports for unit and integration tests will be
reported separately. To this extend, the Codecov configuration is also
updated.