ci: omit coverage for integration for now
The integration testing strategy is less clear than the unit testing
strategy. In most instances where integration coverage is currently
missing this is seemingly because of a lack of integration, thus
everything is already covered by unit tests.

To this extend, run integration tests without coverage in CI and update
the Codecov configuration accordingly.