chore(core,cli): use allowlist-based `.npmignore`
Update the `.npmignore` for the core and cli packages to be allowlist
based. Doing this prevents publishing any files that shouldn't be
published, whether that's version controlled files that we're not added
to the `.npmignore` blocklist or files that happened to exist in the
repository at the time of publishing.