config: convert Stryker configuration to ESM
Rewrite the Stryker configuration as an ESModule from CommonJS. Make
adjacent adjustments in:
- `test.js`: Update the reference to the configuration file.
- `.eslintrc.cjs`: Split JS configuration files override into two, one
  for CJS and one for ESM configuration files, the latter only being
  used for the new Stryker configuration for now. Due to the difference
  in syntax the overridden rules are slightly different.

Also, update the Stryker configuration to not keep it's temporary files
around. This is because, upon success, these are useless (all relevant
information can be found in the report), and upon failure it will be
kept anyway.

Lastly, like some other configuration files, add a comment at the start
with a link to Stryker's homepage.