test(language-css): start `kind` support for CSs value expressions
Rewrite the `cssDeclarationValueExpressionFactory` to start support for
the `kind` field of the `CssDeclarationValueOptions`. This entails
calling the old implementation as a fallback when `kind` isn't set, and
running a (to be implemented) new implementation when it is.

The unit and integration test suites for this factory is updated to, on
top of what it already tested, test the behaviour when `kind` is
defined.