feat(language-css): add `kind`-based implementation for CSS values
Update the CSS value expression factory to work based purely on the
`kind` option of the `CssDeclarationValueOptions`.

The implementation is based entirely off of the pre-existing
implementation, simply delegating calls to the (pre-existing)
`newCssDeclarationValueExpression` function using prefix and suffix
values that come from the mangler plugins (see `language-options.ts`
helpers of mangler plugins).