chore: update license allowlist
Add the "BlueOak-1.0.0" license to the licensee allowlist. This is
another permissive license, most similar to the "BSD-3-Clause" license.
As a permissive license, allowing this is a straightforward decision.

The license is being allowed here due to dependencies updates in [1]
which introduced two new (transitive) dependency that came with this
license, namely:
- jackspeak, v2.2.0
- path-scurry, v1.7.0

--
1. b421932e70f5ccacdc115dc1cd380a7a3ae819dd