config: update ESLint configuration
Update the ESLint configuration file in accordance with dependency
updates to:

- eslint (8.28.0 -> 8.40.0)
- (at)typescript-eslint/eslint-plugin (5.45.0 -> 5.59.2)
- eslint-plugin-import (2.26.0 -> 2.27.5)
- eslint-plugin-jsdoc (39.6.4 -> 44.0.0)
- eslint-plugin-regexp (1.11.0 -> 1.14.0)
- eslint-plugin-security (1.5.0 -> 1.7.1)
- eslint-plugin-yml (1.2.0 -> 1.6.0)

This includes adding newly introduced rules, removing removed rules, and
updating changed rules. This commit _should_ exhaustively cover all base
ESLint rules as well as all (listed) plugin rules.

The rule `jsdoc/informative-docs` is disabled because it introduces new
violations which will not be addressed here because they're too numerous
(count: 63) and not trivial to fix.

The change in the `multi-value-attributes.ts` file is due to the newly
enabled rule `(at)typescript-eslint/lines-around-comment`.

Three new rules for the `(at)typescript-eslint` plugin were explicitly
disabled for markdown snippets because they require a tsconfig.json file
(which isn't present available for snippets).

The rules list for TypesScript snippets in MarkDown is, additionally,
ordered alphabetically as well now.