config: enable new `(at)typescript-eslint` rule
Enable the new rule added in `(at)typescript-eslint/eslint-plugin`
v6.2.0. It's an improvement over a base ESLint rule that takes into
account aspects of the TypeScript language (e.g. explicit/typed
inheritance), so the original rule is explicitly disabled for TypeScript
code (regardless of any configuration elsewhere, we don't want to create
a dependency between the TypeScript specific ruleset and other rulesets
that affect TypeScript code).

The new rule finds violations in two files which are simple enough to be
addressed on demand.