config: update eslint-plugin-regexp configuration
Update the ESLint configuration for the regexp plugin specifically given
changes from 1.15.0 to 2.0.0

All new rules are enabled except `regexp/require-unicode-sets-regexp` to
promote the best as well as consistent regular expressions. The rule
`regexp/require-unicode-sets-regexp` isn't enabled for the same reason
`regexp/require-unicode-regexp` isn't.

Other than that, the major version bump did not appear to break anything
for the usage of the plugin within this project.